lazy output, stop after finding a 100% solution
now Conjure does not test everything before printing.

If an implementation is found,
it is printed right away and the program stops, as it should.

In hindsight, the old behaviour stuck more because of a different algorithm
than an actual need to test everything and sort.  Now the implementation is
evaluated only on the points where the function is defined.

The code became a bit crufty, but this will be fixed in time.