Generate candidates top-down
Due to the use of isNormal instead of isRootNormalE
this implementation is for now much slower,
around 10x for some examples (see diff).

After I change the normalization checking for something more efficient (trie of
Exprs), this difference may become smaller.

Then, I'll be able to start playing tricks that are possible with top-down
generation:

* only allowing ifs at top level
* pattern-matching and cases
* recursion at deconstructed arguments