only recurse inside if bodies
On the factorial example,
this reduces the search space at size 10
from 9065 to 1921 candidates.
4/5 reduction (1/5 of the original candidates).
80% reduction (20% of the original candidates).
However,
runtime only decreases by half:
3.71s to 1.57s.
60% reduction (40% of the original candidates).
3/5 reduction (2/5 of the original candidates).

This is due to the fact that we now have to do multiple enumerations,
one with recursions and one without!  (forN and forR on the diff).

Because of this, runtimes for arith and ints have been increased slightly.

Factorial is the most extreme example of mismatch
between search space and runtime reductions.
For example, "list" had its runtime decrease
proportional to the search space.  (50%)

For the bools example, we now don't see shortcircuiting the || operator for
recursion.  Now recursion appears later through an explicit if.  A fine
compromise.