only allow recursions on one if side
... and require recursions when using ifs

this reduces the search space considerably
from 1/4 down to 2/3 reductions.

fibonacci (size 13)  8567 -> 2377 (/4)
take (size 13)       9766 -> 3618 (/3)
bools (size 10)      1326 ->  464 (/3)
ill-hit (size 10):    193 ->  101 (/2)
ints (size 10)        193 ->  101 (/2)
drop (size 13)        963 ->  601 (2/3)
factorial (size 10)  1389 ->  968 (2/3)

Runtime is not significantly changed.
The reason is that now we have two enumerations
of if applications:
one with recursion on the then clause
the other with recursion on the else clause.

win-lose:
We gain by not having to filter later but
we lose by having to compute a second enumeration.