fix recursion counting in recursiveToDynamic
before recursiveToDynamic was not counting the number of recursions correctly
when processing ifs ands and ors.  If there were recursion on LHSs these were
not counted -- in this case we would even reach infinite loops and get OOMed.

sadly, this fix swaps the output of setelem to an stricter but correct
version...