[BIG] reify instances from each given primitive
For this to happen, conjuring functions had to be changed to take a list of
Prims instead of a list of Exprs ([Prim] becomes [Expr]) and now they must be
provided using the pr and prim functions, equivalent to val and value
functions.

The user boilerplate is sort of the same just with different names.

I could have ultimately kept the same interface by hiding the original "val"
and "value" functions, but this would be confusing in the long term.  Better to
break compatibility earlier than later.

Some runtimes have slighly increased due to the extra "speculating" that has to
happen at the start.

As expected, runtime of bench/lots has increased a lot, as now speculating
about list values is now possible.  (output was changed, but the one recorded
is already the one with the extra instances.)