use the result of given function when possible
when generating the result for a pattern.

For example, if the given partial definition is:

	foo 0  =  1
	foo 1  =  1
	foo 2  =  2

then Conjure will only enumerate the above results themselves for the above
patterns.

There were only two benchmarks with significant performance improvements:
dupos (duplicates and positions) and fibonacci.  On a couple of benchmarks,
there was a small .1 second overhead (arith & bools).

Search space was reduced by roughly half on:

- cube & tnpo from arith.hs
- positions from dupos.hs
- fibonacci.hs (56285 -> 21324)
- length from list.hs (23 -> 13, 25 -> 15)
- set from setelem.hs (466 -> 256)
- tree.hs

Other benchmarks had their search space unchanged or with some unamazing
reductions: factorial was reduced from 343 to 283 and 172 to 148.  On a couple
of occasions, search space has increased.

There is now a dependence between search space and the provided example
function.  An example of this is duplicates and positions:
on duplicates there's no way for this new optimization to "kick-in"
whereas on positions, there is.

Please see the diff of this changeset for more details.