add and use simplifyDefn
Note the generated take and drop functions:

	-- drop 0 []  =  []
	drop 0 (x:xs)  =  x:xs
	drop x []  =  []
	drop x (y:xs)  =  drop (x - 1) xs

	-- take 0 []  =  []
	take 0 (x:xs)  =  []
	take x []  =  []
	take x (y:xs)  =  y:take (x - 1) xs

The first pattern was eliminated.

Perhaps in the future I can update simplifyDefn to make it actually simplify to:

	drop 0 xs  =  xs
	drop x []  =  []
	drop x (y:xs)  =  drop (x - 1) xs

	take 0 xs  =  []
	take x []  =  []
	take x (y:xs)  =  y:take (x - 1) xs

I would have to ensure that all the possibilities the variation of a single
pattern are the same modulo expression matching.

... perhaps reworking the way in which patterns are generated is an easier point of entry.

So that:

	foo 0 xs  = ...
	foo x []  = ...
	foo x (y:xs)  = ...

is actually generated as an option in conjurePats.