prune non-atomic numeric constants
such as:

	1 + 1

	1 + 1 + 1

This can be deactivated with a switch.

One can always include

	2

and

	3

in the list of primitives to arrive at the desired function.

This reduces the number of candidates in the fibonacci example by half.
However runtime is only lowered by 10%.

Other numeric benchmarks had a 10%-ish decrease in runtime...