try to use isRootNormalE again
The most significant effect is in the fibonacci benchmark:

* candidates are cut in four
* but runtime is multiplied by 5!

It is not worth it as it is.

Other benchmarks involving commutative functions also have an increase in
runtime.  Most benchmarks have a little bit of a runtime increase as well.

See the diff for more details.

To be reverted.  Added as a commit to record the effect.