simplify pruning of commutative operators
Before we had:

	isRootNormalC thy (ef :$ ex :$ ey)
	  | ex <= ey  =  True
	  | not (isCommutative thy ef)  =  True
	  | isRootNormal thy (ef :$ ey :$ ex)  =  False

We were only pruning non-ordered commutative expressions
when the ordered version was root-normal.

I thought we could not prune something like the following expression:

	(x - 1) * (x + x)

as its reversal is not normal:

	(x + x) * (x - 1)
	rewrites to
	x * ((x - 1) + (x + 1))
	by
	(x + x) * y  =  x * (y + y)

So then we would have to stick with (x - 1) * (x + x) in the enumeration.
However, x * ((x - 1) + (x + 1)) is equivalent to it
and is bound to appear in the enumeration of expressions of size 9!

So we can definitely prune it,
with the side effect of delaying
some candidates to appear with a bigger size.
This aleady happens anyway, so no harm done.

I now have simply:

	isRootNormalC thy (ef :$ ex :$ ey) | isCommutative thy ef  =  ex <= ey

This can probably be extended to permutative laws in the future.

Runtime has decreased overall as expected:

* fibonacci:  -30%
* carry-on:   -13%
* arith:      -10%
* factorial    -7%
* gps(2):      -7%

Good thing Colin brought this to notice!