update TODO.md
Mergesort is not readily Conjurable.

One can generate takeHalf and dropHalf using Conjure with the following:

	takeHalf :: [Int] -> [Int]
	takeHalf []  =  []
	takeHalf [0,1]  =  [0]
	takeHalf [0,1,2,3]  =  [0,1]

	conjure "takeHalf" takeHalf takeDropIngredients
	  [ fun "take" (take :: Int -> [Int] -> [Int])
	  , fun "drop" (drop :: Int -> [Int] -> [Int])
	  , fun "length" (length :: [Int] -> Int)
	  , fun "`div`" (div :: Int -> Int -> Int)
	  , con (2 :: Int)
	  ]

Then we include [], takeHalf, dropHalf and merge as primitives.

	conjure "mergesort" (sort :: [Int] -> [Int])
	  [ con ([] :: [Int])
	  , fun "takeHalf" (\xs -> take (length xs `div` 2) xs :: [Int])
	  , fun "dropHalf" (\xs -> drop (length xs `div` 2) xs :: [Int])
	  , fun "merge" merge
	  ]

Conjure is not able to find mergesort because of two problems.

The function `takeHalf` is readily identified as a deconstructor on sound
recursions, as it always reduces the length of the list.  However, `dropHalf`
is not: it keeps lists of length 1 with the same length.  We would have to
change the criteria for sound recursions...

The second problem is in the second case of mergesort:

	msort []  =  []
	msort (x:xs)  =  merge (msort (takeHalf (x:xs))) (msort (dropHalf (x:xs)))

We have x:xs, Conjure then does not (currently) allow deconstructing the full
thing (x:xs).  We get xs as arguments and deconstructions of xs.

I am dropping this idea for now...